fix(firmware): PUB byte-count emitted "zu" on target (newlib-nano %zu); build-configurable NATS broker
Verified on hardware: device now connects to NATS over plaintext TCP and publishes
valid Contract B inference messages (PUB edge.line1.cnc-7) the broker accepts.

- nats_proto.c: nats_build_pub used "%zu" for the payload size. The target's newlib-nano
  printf (--specs=nano.specs) lacks the 'z' length modifier, so it emitted the literal
  "zu" as the byte count -> malformed PUB -> broker "Bad or Missing Size" -> connection
  closed. Use %lu with an (unsigned long) cast (portable; identical host output, tests
  still 130/130). Host tests can't catch this — host libc supports %zu.
- device_main.c: NATS_HOST/NATS_PORT are now #ifndef-guarded so they're build-overridable.
- proj_cm33_ns/Makefile: pass through NATS_HOST=, NATS_PORT=, NATS_DISABLE_TLS=1
  (e.g. make build NATS_HOST=10.0.0.5 NATS_DISABLE_TLS=1 for an open LAN broker).

+++ firmware/src/nats_proto.c
@@ -85,7 +85,10 @@ int nats_b64_encode(char *out, size_t cap, const uint8_t *data, size_t len) {
 
 int nats_build_pub(char *buf, size_t cap, const char *subject,
                    const uint8_t *payload, size_t payload_len) {
-    int hn = snprintf(buf, cap, "PUB %s %zu\r\n", subject, payload_len);
+    // %lu + cast, not %zu: the target's newlib-nano printf (--specs=nano.specs) does not
+    // support the 'z' length modifier and would emit the literal "zu" as the byte count,
+    // producing a malformed PUB the broker rejects. unsigned long covers any real payload.
+    int hn = snprintf(buf, cap, "PUB %s %lu\r\n", subject, (unsigned long)payload_len);
     if (hn < 0 || (size_t)hn >= cap) return -1;
     size_t total = (size_t)hn + payload_len + 2;  // payload + trailing CRLF
     if (total > cap) return -1;
